Fixes #2260 Add Initial Conditions Extend and Remove to MoBi.R API
