+++ src/MoBi.Core/Services/AbstractMergeManager.cs
@@ -5,23 +5,8 @@ using OSPSuite.Core.Domain;
 using OSPSuite.Utility.Collections;
 using OSPSuite.Utility.Extensions;
 
-namespace MoBi.Presentation.Tasks
+namespace MoBi.Core.Services
 {
-   public interface IMergeManager<T> where T : class, IObjectBase
-   {
-      /// <summary>
-      ///    Manages the conflict of the two caches
-      /// </summary>
-      /// <param name="merge">The cache being merged into the target</param>
-      /// <param name="target">The target cache of the merge</param>
-      /// <param name="mergeConflictOption">The default option when merging items that are equivalent</param>
-      /// <param name="areElementsEquivalent">
-      ///    Optionally specify a Func that will detect for equivalence between two conflicting
-      ///    elements
-      /// </param>
-      void Merge(ICache<string, T> merge, ICache<string, T> target, MergeConflictOptions mergeConflictOption, Func<T, T, bool> areElementsEquivalent = null);
-   }
-
    public abstract class AbstractMergeManager<T> : IMergeManager<T> where T : class, IObjectBase
    {
       protected Action<T> _removeAction = element => { };
