+++ src/MoBi.Core/Services/NameCorrector.cs
@@ -8,7 +8,7 @@ using OSPSuite.Utility.Extensions;
 
 namespace MoBi.Core.Services
 {
-   public interface INameCorrector
+   public interface INameCorrector : INameAutoCorrector
    {
       /// <summary>
       ///    Modifies a name on a named object if its name is already part of the enumeration
@@ -36,32 +36,18 @@ namespace MoBi.Core.Services
       /// <returns>true if a modification was made, otherwise false</returns>
       bool CorrectName<T>(IEnumerable<string> alreadyUsedNames, T objectForRename) where T : IObjectBase;
 
-      /// <summary>
-      ///    Modifies a name on a named object if its name is already part of the enumeration
-      ///    <paramref name="alreadyUsedNames" />.
-      ///    User confirmation of name change is not required
-      /// </summary>
-      /// <param name="alreadyUsedNames">
-      ///    The list of names that are already used. These names cannot be considered for a name
-      ///    correction
-      /// </param>
-      /// <param name="objectForRename">The object whose name is being corrected</param>
-      void AutoCorrectName<T>(IEnumerable<string> alreadyUsedNames, T objectForRename) where T : IObjectBase;
-
       string PromptForCorrectName<T>(IReadOnlyList<string> alreadyUsedNames, T objectForRename) where T : IObjectBase;
    }
 
-   internal class NameCorrector : INameCorrector
+   internal class NameCorrector : NameAutoCorrector, INameCorrector
    {
       private readonly IObjectBaseNamingTask _namingTask;
       private readonly IObjectTypeResolver _objectTypeResolver;
-      private readonly IContainerTask _containerTask;
 
-      public NameCorrector(IObjectBaseNamingTask namingTask, IObjectTypeResolver objectTypeResolver, IContainerTask containerTask)
+      public NameCorrector(IObjectBaseNamingTask namingTask, IObjectTypeResolver objectTypeResolver, IContainerTask containerTask) : base(containerTask)
       {
          _namingTask = namingTask;
          _objectTypeResolver = objectTypeResolver;
-         _containerTask = containerTask;
       }
 
       public bool CorrectName<T>(IEnumerable<T> alreadyExistingNamedObjects, T objectForRename) where T : IObjectBase
@@ -95,25 +81,12 @@ namespace MoBi.Core.Services
             newName = _namingTask.NewName(
                AppConstants.Dialog.AskForChangedName(oldName, typeName),
                Captions.Rename,
-               getNextSuggestedName(usedNames, oldName),
+               GetNextSuggestedName(usedNames, oldName),
                usedNames,
                iconName: ApplicationIcons.Rename.IconName);
          }
 
          return newName;
-      }
-
-      public void AutoCorrectName<T>(IEnumerable<string> alreadyUsedNames, T objectForRename) where T : IObjectBase
-      {
-         var oldName = objectForRename.Name;
-         var updatedName = getNextSuggestedName(alreadyUsedNames, oldName, canUseBaseName: true);
-
-         objectForRename.Name = updatedName;
-      }
-
-      private string getNextSuggestedName(IEnumerable<string> usedNames, string oldName, bool canUseBaseName = false)
-      {
-         return _containerTask.CreateUniqueName(usedNames.ToList(), oldName, canUseBaseName);
       }
    }
 }
