+++ src/MoBi.Presentation/Presenter/EditFormulaInPathAndValuesPresenter.cs
@@ -1,7 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using MoBi.Assets;
 using MoBi.Core.Domain.Model;
 using MoBi.Core.Domain.Services;
 using MoBi.Core.Helper;
+using MoBi.Core.Services;
 using MoBi.Presentation.Mappers;
 using MoBi.Presentation.Views;
 using OSPSuite.Core.Domain;
@@ -21,6 +25,8 @@ namespace MoBi.Presentation.Presenter
 
    public class EditFormulaInPathAndValuesPresenter : EditFormulaPresenter<IEditFormulaInPathAndValuesView, IEditFormulaInPathAndValuesPresenter>, IEditFormulaInPathAndValuesPresenter
    {
+      private readonly IObjectBaseNamingTask _namingTask;
+
       public EditFormulaInPathAndValuesPresenter(IEditFormulaInPathAndValuesView view,
          IFormulaPresenterCache formulaPresenterCache,
          IMoBiContext context,
@@ -28,20 +34,29 @@ namespace MoBi.Presentation.Presenter
          FormulaTypeCaptionRepository formulaTypeCaptionRepository,
          IMoBiFormulaTask formulaTask,
          ICircularReferenceChecker circularReferenceChecker,
-         ISelectReferenceAtParameterValuePresenter referencePresenter) : base(view, formulaPresenterCache, context, formulaDTOMapper, formulaTask, formulaTypeCaptionRepository, circularReferenceChecker)
+         ISelectReferenceAtParameterValuePresenter referencePresenter, IObjectBaseNamingTask namingTask) : base(view, formulaPresenterCache, context, formulaDTOMapper, formulaTask, formulaTypeCaptionRepository, circularReferenceChecker)
       {
+         _namingTask = namingTask;
          ReferencePresenter = referencePresenter;
       }
 
       public void AddNewFormula(string formulaName = null)
       {
-         var (_, newFormula) = _formulaTask.CreateNewFormulaInBuildingBlock(_formulaDTO.Type, FormulaDimension, AllFormulaNames, _buildingBlock, formulaName);
+         if (shouldNameFormula(formulaName, AllFormulaNames.ToList()))
+            formulaName = _namingTask.NewName(AppConstants.Captions.NewName, AppConstants.Captions.EnterNewFormulaName, string.Empty, AllFormulaNames);
+
+         if (string.IsNullOrEmpty(formulaName))
+            return;
+
+         var (_, newFormula) = _formulaTask.CreateNewFormulaInBuildingBlock(_formulaDTO.Type, FormulaDimension, _buildingBlock, formulaName);
          if (newFormula == null)
             return;
 
          SelectFormula(newFormula);
          UpdateFormula();
       }
+
+      private static bool shouldNameFormula(string newFormulaName, IReadOnlyList<string> forbiddenNames) => string.IsNullOrEmpty(newFormulaName) || forbiddenNames.Contains(newFormulaName);
 
       public void Init<TBuildingBlock, TBuilder>(TBuilder formulaOwner, TBuildingBlock buildingBlock, UsingFormulaDecoder formulaDecoder) where TBuilder : PathAndValueEntity, IUsingFormula, IWithDisplayUnit where TBuildingBlock : class, ILookupBuildingBlock<TBuilder>
       {
