--- src/MoBi.Presentation/Tasks/InitialConditionsBuildingBlockExtendManager.cs
@@ -1,12 +1,0 @@
-﻿using OSPSuite.Core.Domain.Builder;
-
-namespace MoBi.Presentation.Tasks
-{
-   public interface IInitialConditionsBuildingBlockExtendManager : IExtendPathAndValuesManager<InitialCondition>
-   {
-   }
-
-   public class InitialConditionsBuildingBlockExtendManager : ExtendPathAndValuesManager<InitialCondition>, IInitialConditionsBuildingBlockExtendManager
-   {
-   }
-}
