+++ src/MoBi.R/Services/InitialConditionsTask.cs
@@ -1,0 +1,59 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using MoBi.Assets;
+using MoBi.Core.Commands;
+using MoBi.Core.Domain.Model;
+using MoBi.Core.Extensions;
+using MoBi.Core.Services;
+using OSPSuite.Core.Domain.Builder;
+using OSPSuite.Core.Domain.Services;
+using OSPSuite.Utility.Extensions;
+
+namespace MoBi.R.Services;
+
+public interface IInitialConditionsTask
+{
+   void DeleteInitialConditions(InitialConditionsBuildingBlock buildingBlock, string[] pathsToDelete);
+
+   void ExtendInitialConditions(InitialConditionsBuildingBlock buildingBlock, MoBiSpatialStructure spatialStructure, MoleculeBuildingBlock moleculeBuildingBlock, string[] moleculeNames);
+}
+
+public class InitialConditionsTask : IInitialConditionsTask
+{
+   private readonly IMoBiContext _context;
+   private readonly IObjectTypeResolver _objectTypeResolver;
+   private readonly IInitialConditionsBuildingBlockExtendManager _extendManager;
+
+   public InitialConditionsTask(IMoBiContext context, IObjectTypeResolver objectTypeResolver, IInitialConditionsBuildingBlockExtendManager extendManager)
+   {
+      _context = context;
+      _objectTypeResolver = objectTypeResolver;
+      _extendManager = extendManager;
+   }
+
+   public void DeleteInitialConditions(InitialConditionsBuildingBlock buildingBlock, string[] pathsToDelete)
+   {
+      var initialConditionsToDelete = pathsToDelete.Select(buildingBlock.FindByPath).Where(x => x != null).ToList();
+      var macroCommand = new MoBiMacroCommand
+      {
+         Description = AppConstants.Commands.RemoveMultipleInitialConditions,
+         CommandType = AppConstants.Commands.DeleteCommand,
+         ObjectType = _objectTypeResolver.TypeFor<InitialCondition>()
+      };
+
+      macroCommand.AddRange(initialConditionsToDelete.Select(ic => new RemoveInitialConditionFromBuildingBlockCommand(buildingBlock, ic.Path)));
+
+      _context.AddToHistory(macroCommand.RunCommand(_context));
+   }
+
+   public void ExtendInitialConditions(InitialConditionsBuildingBlock buildingBlock, MoBiSpatialStructure spatialStructure, MoleculeBuildingBlock moleculeBuildingBlock, string[] moleculeNames)
+   {
+      IReadOnlyList<MoleculeBuilder> molecules;
+      if (moleculeNames == null || !moleculeNames.Any())
+         molecules = moleculeBuildingBlock.All().ToList();
+      else
+         molecules = moleculeBuildingBlock.Where(x => moleculeNames.Contains(x.Name)).ToList();
+
+      _context.AddToHistory(_extendManager.ExtendPathAndValueEntitiesBasedOnUsedTemplates(spatialStructure, molecules, buildingBlock));
+   }
+}
